pyproject.toml
[project]
name = "websum-to-git"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "tiaot33", email = "lyq12123@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "requests",
    "pyyaml",
    "beautifulsoup4",
    "python-telegram-bot[job-queue]",
    "openai",
    "anthropic",
    "google-genai",
    "readability-lxml",
    "markdownify",
    "tiktoken",
    "camoufox[geoip]",
    "PyGithub",
]

[project.scripts]
web-sum-to-git = "websum_to_git:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "ruff>=0.14.6",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=6.0.0",
    "responses>=0.25.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-v --tb=short"
markers = [
    "functional: 需要外部服务（Telegram/LLM/GitHub 等）的真实功能测试",
]

[tool.ruff]
# Set the maximum line length to 120.
line-length = 120
[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "W",   # pycodestyle
    "F",   # Pyflakes
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "I",   # isort
]

[tool.uv.sources]
camoufox = { git = "https://github.com/coryking/camoufox.git", subdirectory = "pythonlib", rev = "v142.0.1-fork.27" }
